pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "jobproc"
version = "0.1.0"
description = "Job-listing ingestion + embedding neighborhood analysis"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Alex Ilin"}]
license = {text = "MIT"}
dependencies = [
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "scipy>=1.11",
    "psycopg2-binary>=2.9",
    "pgvector>=0.2.5",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "google-genai>=0.3",
]

[project.optional-dependencies]
dev = ["pytest>=7.4"]

[tool.setuptools.packages.find]
include = ["jobproc*"]
